Upgrade to Storybook 7
