+++ package.json
@@ -19,13 +19,13 @@
     "patch-package": "patch-package",
     "build": "yarn build-images && yarn patch-package && babel src -d dist --extensions \".js,.jsx,.ts,.tsx\" --ignore \"**/*.stories.tsx\" --ignore \"**/*.stories.js\"",
     "build-images": "yarn svgr --out-dir src/images src/images",
-    "build-storybook": "yarn build-images && yarn patch-package && build-storybook -s .storybook/static",
+    "build-storybook": "yarn build-images && yarn patch-package && storybook build -s .storybook/static",
     "clean": "rm -rf dist",
     "lint": "yarn lint:js && yarn lint:package",
     "lint:js": "cross-env NODE_ENV=production eslint --cache --cache-location=.cache/eslint --ext .js,.jsx,.html,.ts,.tsx,.mjs --report-unused-disable-directives",
     "lint:package": "sort-package-json",
     "release": "dotenv yarn build & yarn typescript:generate && auto shipit",
-    "storybook": "yarn patch-package && start-storybook -p 6006 -s .storybook/static",
+    "storybook": "yarn patch-package && storybook dev -p 6006 -s .storybook/static",
     "typescript:check": "tsc --project ./tsconfig.json --noEmit",
     "typescript:generate": "tsc --declaration --emitDeclarationOnly --outDir dist --declarationMap"
   },
@@ -50,7 +50,7 @@
   },
   "dependencies": {
     "@emotion/weak-memoize": "^0.2.5",
-    "@storybook/theming": "^6.5.9",
+    "@storybook/theming": "^7.0.0-rc.1",
     "@types/pluralize": "^0.0.29",
     "@types/prismjs": "^1.16.6",
     "@types/react-modal": "^3.12.1",
@@ -75,14 +75,15 @@
     "@babel/preset-react": "^7.10.1",
     "@babel/preset-typescript": "^7.10.1",
     "@emotion/babel-plugin": "^11.9.2",
-    "@storybook/addon-a11y": "^6.5.9",
-    "@storybook/addon-actions": "^6.5.9",
-    "@storybook/addon-docs": "^6.5.9",
-    "@storybook/addon-essentials": "^6.5.9",
-    "@storybook/addon-storysource": "^6.5.9",
+    "@storybook/addon-a11y": "^7.0.0-rc.1",
+    "@storybook/addon-actions": "^7.0.0-rc.1",
+    "@storybook/addon-docs": "^7.0.0-rc.1",
+    "@storybook/addon-essentials": "^7.0.0-rc.1",
+    "@storybook/addon-storysource": "^7.0.0-rc.1",
     "@storybook/eslint-config-storybook": "^3.1.2",
     "@storybook/linter-config": "^3.1.2",
-    "@storybook/react": "^6.5.9",
+    "@storybook/react": "^7.0.0-rc.1",
+    "@storybook/react-webpack5": "^7.0.0-rc.1",
     "@svgr/cli": "^6.1.1",
     "@types/fs-extra": "^9.0.1",
     "auto": "^9.50.1",
@@ -92,7 +93,7 @@
     "cross-env": "^7.0.2",
     "dotenv-cli": "^3.1.0",
     "eslint": "^7.1.0",
-    "eslint-plugin-storybook": "^0.5.12",
+    "eslint-plugin-storybook": "^0.6.11",
     "husky": "^4.2.5",
     "lint-staged": "^10.2.9",
     "node-fetch": "^2.6.0",
@@ -102,8 +103,9 @@
     "react-dom": "17",
     "seedrandom": "^3.0.5",
     "sort-package-json": "^1.51.0",
+    "storybook": "^7.0.0-rc.1",
     "ts-loader": "^7.0.5",
-    "typescript": "^3.9.5"
+    "typescript": "^4.9.5"
   },
   "peerDependencies": {
     "react": "^15.0.0 || ^16.0.0 || ^17.0.0",
