+++ src/Intro.mdx
@@ -1,4 +1,4 @@
-import { Meta } from '@storybook/addon-docs/';
+import { Meta } from '@storybook/blocks';
 import designSystem from './design-system.png';
 
 <Meta title="Intro" />
