+++ src/components/Color.mdx
@@ -1,4 +1,4 @@
-import { Meta, ColorPalette, ColorItem } from '@storybook/addon-docs/';
+import { Meta, ColorPalette, ColorItem } from '@storybook/blocks';
 
 import { color } from './shared/styles';
