+++ src/components/Typography.mdx
@@ -1,4 +1,4 @@
-import { Meta, Typeset } from '@storybook/addon-docs/';
+import { Meta, Typeset } from '@storybook/blocks';
 import { typography } from './shared/styles';
 
 <Meta title="Typography/Base" />
